backend/app/services/model_service.py
# services/model_service.py
import torch
import torch.nn as nn
import numpy as np
from pathlib import Path
import joblib
import logging
from typing import Dict, List, Any, Optional
from datetime import datetime

logger = logging.getLogger(__name__)

class ModelService:
    def __init__(self):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.models_dir = Path("ml_engine/models")
        self.models_dir.mkdir(parents=True, exist_ok=True)
        self.model_loaded = False
        self.model_info = {}
        logger.info(f"Using device: {self.device}")

    def load_model(self):
        """Load the trained model"""
        try:
            # For now, we'll create a simple model since the GAN might not be trained yet
            self.model_loaded = True
            self.model_info = {
                'model_type': 'UrbanGrowthPredictor',
                'features_count': 8,
                'training_date': datetime.now().isoformat(),
                'accuracy': 0.85,
                'last_retrained': datetime.now().isoformat()
            }
            logger.info("✅ Model service initialized (mock mode)")
            return True
        except Exception as e:
            logger.error(f"Failed to load model: {e}")
            self.model_loaded = False
            return False

    def is_loaded(self):
        """Check if model is loaded"""
        return self.model_loaded

    def get_model_info(self):
        """Get model information"""
        return self.model_info

    def predict(self, features: List[List[float]]) -> List[float]:
        """Make predictions using the model"""
        try:
            if not self.model_loaded:
                # Return mock predictions if model not loaded
                return [np.random.uniform(0.02, 0.15) for _ in features]
            
            # Simple linear model for demonstration
            predictions = []
            for feature_set in features:
                if len(feature_set) >= 8:
                    # Mock prediction logic - replace with actual model
                    base_growth = feature_set[0] * 0.1 + feature_set[4] * 0.3
                    climate_factor = feature_set[5] * 0.2
                    prediction = max(0.01, min(0.25, base_growth + climate_factor + np.random.normal(0, 0.02)))
                    predictions.append(prediction)
                else:
                    predictions.append(0.05)  # Default growth rate
            
            return predictions
        except Exception as e:
            logger.error(f"Prediction failed: {e}")
            return [0.05] * len(features)  # Return default predictions

    def get_feature_importance(self) -> Dict[str, float]:
        """Get feature importance scores"""
        return {
            'urban_density': 0.25,
            'vegetation_cover': 0.15,
            'built_up_area': 0.20,
            'urban_std': 0.10,
            'historical_growth': 0.18,
            'climate_scenario': 0.12
        }

    def retrain_model(self, training_data_path: Optional[str] = None, epochs: int = 100, force_retrain: bool = False):
        """Retrain the model (mock implementation)"""
        try:
            logger.info(f"🔄 Starting model retraining (epochs: {epochs})")
            
            # Simulate training process
            import time
            time.sleep(2)  # Simulate training time
            
            # Update model info
            self.model_info.update({
                'training_date': datetime.now().isoformat(),
                'last_retrained': datetime.now().isoformat(),
                'accuracy': min(0.95, self.model_info.get('accuracy', 0.85) + 0.02)
            })
            
            logger.info("✅ Model retraining completed")
            return True
        except Exception as e:
            logger.error(f"Model retraining failed: {e}")
            return False

    def generate_future_scenarios(self, climate_data: Dict, num_scenarios: int = 4) -> List[Dict[str, Any]]:
        """Generate future urban scenarios (mock implementation)"""
        try:
            scenarios = []
            
            scenario_types = [
                "Climate-Resilient Development",
                "High-Density Urban Expansion", 
                "Sustainable Green Growth",
                "Mixed-Use Balanced Development"
            ]
            
            for i, scenario_type in enumerate(scenario_types[:num_scenarios]):
                scenario = {
                    'id': i + 1,
                    'name': scenario_type,
                    'description': f"{scenario_type} scenario for future urban planning",
                    'urban_growth_rate': np.random.uniform(0.08, 0.25),
                    'vegetation_change': np.random.uniform(-0.1, 0.2),
                    'climate_resilience': np.random.uniform(0.6, 0.95),
                    'infrastructure_density': np.random.uniform(0.3, 0.9),
                    'key_features': [
                        "Smart urban planning",
                        "Climate adaptation",
                        "Sustainable infrastructure"
                    ]
                }
                scenarios.append(scenario)
            
            logger.info(f"Generated {len(scenarios)} urban scenarios")
            return scenarios
            
        except Exception as e:
            logger.error(f"Scenario generation failed: {e}")
            return []

# Singleton instance
model_service = ModelService()