backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
earthengine-api==0.1.400
geemap==0.29.3
geopy==2.4.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.2
joblib==1.3.2
folium==0.15.1
rasterio==1.3.9
pillow==10.1.0
python-multipart==0.0.6
flask==2.3.3
flask-cors==4.0.0
# ... keep the rest of your dependencies